File Object

Represents a file output from a workflow. A File has an absolute path and a basename. Given a suffix pattern, it can also have a prefix (e.g. for file foo.txt and a \\.txt suffix pattern, the prefix will be foo).

Public fields

path

(character(1))
Absolute path of file.

bname

(character(1))
Basename of file.

suffix_pattern

(character(1))
Suffix pattern.

size

(character(1))
Size of file.

prefix

(character(1))
Prefix of file.

Methods

Public methods

File$new()

File$print()

File$clone()

Method new()

Create a new File object.

Usage

File$new(path = NULL, suffix_pattern = NULL)

Arguments

path

(character(1))
Absolute path of file.

suffix_pattern

(character(1))
Suffix pattern.

Method print()

Print details about the File.

Usage

File$print(...)

Arguments

...

(ignored).

Method clone()

The objects of this class are cloneable with this method.

Usage

File$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

f1 <- File$new(readr::readr_example("mtcars.csv"), "\\.csv")
(b <- f1$bname)
#> [1] "mtcars.csv"
(sp <- f1$suffix_pattern)
#> [1] "\\.csv"
(p <- f1$prefix)
#> [1] "mtcars"