Virusinterpreter Object

Virusinterpreter file parsing and manipulation.

Super class

tidywigits::Tool -> Virusinterpreter

Methods

Public methods

Virusinterpreter$new()

Virusinterpreter$parse_annotated()

Virusinterpreter$tidy_annotated()

Virusinterpreter$clone()

Inherited methods

tidywigits::Tool$.eval_func()

tidywigits::Tool$.parse_file()

tidywigits::Tool$.parse_file_nohead()

tidywigits::Tool$.tidy_file()

tidywigits::Tool$filter_files()

tidywigits::Tool$list_files()

tidywigits::Tool$nemofy()

tidywigits::Tool$print()

tidywigits::Tool$tidy()

tidywigits::Tool$write()

Method new()

Create a new Virusinterpreter object.

Usage

Virusinterpreter$new(
  path = NULL,
  files_tbl = NULL,
  tidy = TRUE,
  keep_raw = FALSE
)

Arguments

path

(character(1))
Output directory of tool. If files_tbl is supplied, this basically gets ignored.

files_tbl

(tibble(n))
Tibble of files from list_files_dir().

tidy

(logical(1))
Should the raw parsed tibbles get tidied?

keep_raw

(logical(1))
Should the raw parsed tibbles be kept in the final output?

Method parse_annotated()

Read virus.annotated.tsv file.

Usage

Virusinterpreter$parse_annotated(x)

Arguments

x

(character(1))
Path to file.

Method tidy_annotated()

Tidy virus.annotated.tsv file.

Usage

Virusinterpreter$tidy_annotated(x)

Arguments

x

(character(1))
Path to file.

Method clone()

The objects of this class are cloneable with this method.

Usage

Virusinterpreter$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

cls <- Virusinterpreter
indir <- system.file("extdata/oa", package = "tidywigits")
odir <- tempdir()
id <- "virusinterpreter_run1"
obj <- cls$new(indir)
obj$nemofy(odir = odir, format = "parquet", id = id)
list.files(odir, pattern = "parquet", full.names = FALSE)
#>  [1] "sample1_2_sage_bqrtsv.parquet"                
#>  [2] "sample1_amber_bafpcf.parquet"                 
#>  [3] "sample1_amber_contaminationtsv.parquet"       
#>  [4] "sample1_amber_homozygousregion.parquet"       
#>  [5] "sample1_amber_qc.parquet"                     
#>  [6] "sample1_chord_prediction.parquet"             
#>  [7] "sample1_chord_signatures.parquet"             
#>  [8] "sample1_cobalt_gcmed_bucket_stats.parquet"    
#>  [9] "sample1_cobalt_gcmed_sample_stats.parquet"    
#> [10] "sample1_cobalt_ratiomed.parquet"              
#> [11] "sample1_cobalt_ratiopcf.parquet"              
#> [12] "sample1_cuppa_datatsv.parquet"                
#> [13] "sample1_cuppa_feat.parquet"                   
#> [14] "sample1_cuppa_predsum.parquet"                
#> [15] "sample1_germline_linx_breakends.parquet"      
#> [16] "sample1_germline_linx_clusters.parquet"       
#> [17] "sample1_germline_linx_drivercatalog.parquet"  
#> [18] "sample1_germline_linx_links.parquet"          
#> [19] "sample1_germline_linx_svs.parquet"            
#> [20] "sample1_germline_purple_drivercatalog.parquet"
#> [21] "sample1_lilac_qc.parquet"                     
#> [22] "sample1_lilac_summary.parquet"                
#> [23] "sample1_linx_breakends.parquet"               
#> [24] "sample1_linx_clusters.parquet"                
#> [25] "sample1_linx_drivercatalog.parquet"           
#> [26] "sample1_linx_drivers.parquet"                 
#> [27] "sample1_linx_fusions.parquet"                 
#> [28] "sample1_linx_links.parquet"                   
#> [29] "sample1_linx_svs.parquet"                     
#> [30] "sample1_linx_viscn.parquet"                   
#> [31] "sample1_linx_visfusion.parquet"               
#> [32] "sample1_linx_visgeneexon.parquet"             
#> [33] "sample1_linx_visproteindomain.parquet"        
#> [34] "sample1_linx_vissegments.parquet"             
#> [35] "sample1_linx_vissvdata.parquet"               
#> [36] "sample1_purple_cnvgenetsv.parquet"            
#> [37] "sample1_purple_cnvsomtsv.parquet"             
#> [38] "sample1_purple_drivercatalog.parquet"         
#> [39] "sample1_purple_germdeltsv.parquet"            
#> [40] "sample1_purple_purityrange.parquet"           
#> [41] "sample1_purple_puritytsv.parquet"             
#> [42] "sample1_purple_qc.parquet"                    
#> [43] "sample1_purple_somclonality.parquet"          
#> [44] "sample1_purple_somhist.parquet"               
#> [45] "sample1_sage_bqrtsv.parquet"                  
#> [46] "sample1_sage_exoncvg.parquet"                 
#> [47] "sample1_sage_genecvg_cvg.parquet"             
#> [48] "sample1_sage_genecvg_genes.parquet"           
#> [49] "sample1_sigs_allocation.parquet"              
#> [50] "sample1_sigs_snvcounts.parquet"               
#> [51] "sample1_virusbreakend_summary.parquet"        
#> [52] "sample1_virusinterpreter_annotated.parquet"   
#> [53] "sample2_2_sage_bqrtsv.parquet"                
#> [54] "sample2_sage_bqrtsv.parquet"                  
#> [55] "sample2_sage_exoncvg.parquet"                 
#> [56] "sample2_sage_genecvg_cvg.parquet"             
#> [57] "sample2_sage_genecvg_genes.parquet"           
#> [58] "version_2_linx_version.parquet"               
#> [59] "version_cobalt_version.parquet"               
#> [60] "version_linx_version.parquet"                 
#> [61] "version_purple_version.parquet"