Virusinterpreter Object

Virusinterpreter file parsing and manipulation.

Super class

tidywigits::Tool -> Virusinterpreter

Methods

Public methods

Virusinterpreter$new()

Virusinterpreter$parse_annotated()

Virusinterpreter$tidy_annotated()

Virusinterpreter$clone()

Inherited methods

tidywigits::Tool$.eval_func()

tidywigits::Tool$.parse_file()

tidywigits::Tool$.parse_file_nohead()

tidywigits::Tool$.tidy_file()

tidywigits::Tool$filter_files()

tidywigits::Tool$list_files()

tidywigits::Tool$nemofy()

tidywigits::Tool$print()

tidywigits::Tool$tidy()

tidywigits::Tool$write()

Method new()

Create a new Virusinterpreter object.

Usage

Virusinterpreter$new(
  path = NULL,
  files_tbl = NULL,
  tidy = TRUE,
  keep_raw = FALSE
)

Arguments

path

(character(1))
Output directory of tool. If files_tbl is supplied, this basically gets ignored.

files_tbl

(tibble(n))
Tibble of files from list_files_dir.

tidy

(logical(1))
Should the raw parsed tibbles get tidied?

keep_raw

(logical(1))
Should the raw parsed tibbles be kept in the final output?

Method parse_annotated()

Read virus.annotated.tsv file.

Usage

Virusinterpreter$parse_annotated(x)

Arguments

x

(character(1))
Path to file.

Method tidy_annotated()

Tidy virus.annotated.tsv file.

Usage

Virusinterpreter$tidy_annotated(x)

Arguments

x

(character(1))
Path to file.

Method clone()

The objects of this class are cloneable with this method.

Usage

Virusinterpreter$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

if (FALSE) { # \dontrun{
path <- here::here("nogit")
v <- Virusinterpreter$new(path)
} # }